+++ EntitiesEnhancerSourceGenerator/EntitiesEnhancerSourceGenerator/EntitiesEnhancerSourceGenerator.Sample/UnityMock.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using Unity.Collections;
+﻿using Unity.Collections;
 
 namespace Unity.Entities
 {
