+++ EntitiesEnhancerSourceGenerator/EntitiesEnhancerSourceGenerator/EntitiesEnhancerSourceGenerator/EntitiesEnhancerSourceGenerator.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using EntityEnhancerSourceGenerator;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
